requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
torchaudio
openai-whisper
pyannote.audio
